Configure colorcolumn and ruler to remember to keep lines and files short
The colorcolumn is a new feature introduced in vim 7.3, which highlights
one or several columns at given positions. I configured the position at
81 characters to keep the highlight out of the view in a typical terminal,
making its display a warning against any line stepping on the line.

The ruler option displays the line and column of the cursor position,
and an indication of the position in the file: Top, Bottom, All, or
a percentage relative to the total length of the file. I find the
constant update of display in the ruler distracting, and thus chose
to display only the current line number and the total, in the following
format: current/total, e.g. 34/220. This provides a reminder to keep
files short, typically between 100 and 200 lines.