Allow the cursor everywhere in the screen in all modes
This avoids having the cursor moving left and right to get to the
last character of a line when browsing. Also, it makes it possible
to create sparse tables more easily, e.g. column by column. When
editing at a virtual position, spaces are added to fill the start
of the line up to the current position.

Kudos to Vladimir Keleshev for sharing this tip in:

  Writing a Book with Pandoc, Make, and Vim
  2020-04-12, by Vladimir Keleshev
  https://keleshev.com/my-book-writing-setup/